:recycle: separate unit test and integration test
