+++ ozkour-back/package.json
@@ -5,8 +5,10 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "npm run test:unit && npm run lint",
-    "test:unit": "jest",
+    "test": "npm run test:all && npm run lint",
+    "test:all": "jest",
+    "test:unit": "jest tests/unit",
+    "test:integration": "jest tests/integration",
     "dev": "node ./src",
     "dev:watch": "nodemon ./src",
     "lint": "eslint --ext .js .",
