+++ ozkour-back/tests/unit/slide.test.js
@@ -1,7 +1,5 @@
-const googleSlideRepository = require('../src/infrastructure/googleslide/googleSlideRepository')
-const { presentationId, getSlides } = require('../src/infrastructure/googleslide/slideWrapper')
-const { SlideService } = require('../src/domain/slideService')
-const connect = require('../src/infrastructure/connect.js')
+const googleSlideRepository = require('../../src/infrastructure/googleslide/googleSlideRepository')
+const { SlideService } = require('../../src/domain/slideService')
 
 const slideServiceRepository = googleSlideRepository
 const slideService = new SlideService(slideServiceRepository)
@@ -130,35 +128,6 @@ describe('Slides creation', () => {
     const registerSpeakers = googleSlideRepository.addSpeakersWithStyleToTable(dateId, speakers, IndexRowInTableToInsert)
     // then
     expect(registerSpeakers).toMatchSnapshot()
-  })
-})
-
-describe('Integration test on create a slide', () => {
-  it('should generate expected Google Slide file', async () => {
-    // given
-    const talks = [_createValidTalk()]
-    try {
-      // when
-      connect.createAuth()
-      const res = await slideService.createSlides(talks)
-      const slides = await getSlides()
-      // then
-      expect(res).toStrictEqual({
-        message: 'Created !',
-        link: 'https://docs.google.com/presentation/d/' +
-            presentationId +
-            '/'
-      })
-      expect(JSON.stringify(slides[1])
-        .replace(/"objectId":".*?",/g, '"objectId":"id",')
-        .replace(/"speakerNotesObjectId":".*?"/g, '"speakerNotesObjectId":"id"')
-        .replace(/"https:\/\/lh[1-9].googleusercontent.com\/.*?",/g, '"lien",')
-        .replace(/"listId":".*?"/g, '"listId":"listId"')
-        .replace(/"lists":{".*?"/g, '"lists":"lists"')).toMatchSnapshot()
-    } catch (e) {
-      console.error(e)
-      throw e
-    }
   })
 })
